# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Expense Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  upiId?: string | null;
  avatarColor: string;
  createdAt: string;
}

export interface CreateUserBody {
  name: string;
  email: string;
  /** @nullable */
  upiId?: string | null;
  avatarColor: string;
}

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  emoji: string;
  createdByUserId: number;
  createdAt: string;
}

export interface GroupDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  emoji: string;
  createdByUserId: number;
  createdAt: string;
  members: User[];
  totalExpenses: number;
  expenseCount: number;
}

export interface CreateGroupBody {
  name: string;
  /** @nullable */
  description?: string | null;
  emoji: string;
  createdByUserId: number;
  memberUserIds?: number[];
}

export interface UpdateGroupBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  emoji?: string;
}

export interface GroupMember {
  id: number;
  groupId: number;
  userId: number;
  joinedAt: string;
}

export interface AddGroupMemberBody {
  userId: number;
}

export interface ExpenseSplit {
  id: number;
  expenseId: number;
  userId: number;
  userName: string;
  amount: number;
  isPaid: boolean;
}

export interface ExpenseWithSplits {
  id: number;
  groupId: number;
  groupName: string;
  description: string;
  amount: number;
  category: string;
  paidByUserId: number;
  paidByUserName: string;
  createdAt: string;
  splits: ExpenseSplit[];
}

export type CreateExpenseBodySplitsItem = {
  userId: number;
  amount: number;
};

export interface CreateExpenseBody {
  groupId: number;
  description: string;
  amount: number;
  category: string;
  paidByUserId: number;
  splits: CreateExpenseBodySplitsItem[];
}

export interface UpdateExpenseBody {
  requestingUserId: number;
  description?: string;
  amount?: number;
  category?: string;
}

export interface DeleteExpenseBody {
  requestingUserId: number;
}

export interface Settlement {
  id: number;
  groupId: number;
  fromUserId: number;
  fromUserName: string;
  toUserId: number;
  toUserName: string;
  /** @nullable */
  toUserUpiId?: string | null;
  amount: number;
  /** @nullable */
  upiLink?: string | null;
  settledAt: string;
}

export interface CreateSettlementBody {
  groupId: number;
  fromUserId: number;
  toUserId: number;
  amount: number;
}

export type BalanceDirection =
  (typeof BalanceDirection)[keyof typeof BalanceDirection];

export const BalanceDirection = {
  owe: "owe",
  owed: "owed",
} as const;

export interface Balance {
  groupId: number;
  groupName: string;
  groupEmoji: string;
  withUserId: number;
  withUserName: string;
  /** @nullable */
  withUserUpiId?: string | null;
  amount: number;
  direction: BalanceDirection;
}

export interface DashboardSummary {
  totalOwed: number;
  totalOwe: number;
  netBalance: number;
  groupCount: number;
  recentExpenseCount: number;
  settledThisMonth: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  expense_added: "expense_added",
  settlement_done: "settlement_done",
  group_joined: "group_joined",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  amount?: number | null;
  groupName: string;
  actorName: string;
  createdAt: string;
}

export type ListExpensesParams = {
  groupId?: number;
};

export type ListSettlementsParams = {
  groupId?: number;
};

export type GetDashboardSummaryParams = {
  userId: number;
};

export type GetUserBalancesParams = {
  userId: number;
};

export type GetRecentActivityParams = {
  userId: number;
};