# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pulse Expense Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses, optionally filtered by group
      parameters:
        - name: groupId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseWithSplits"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithSplits"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithSplits"
        "404":
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseWithSplits"
        "403":
          description: Forbidden - not the owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteExpenseBody"
      responses:
        "204":
          description: Expense deleted
        "403":
          description: Forbidden - not the owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /dashboard/balances:
    get:
      operationId: getUserBalances
      tags: [dashboard]
      summary: Get all balances for a user across all groups
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Balance"
  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ExpenseSplit:
      type: object
      properties:
        id:
          type: integer
        expenseId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        isPaid:
          type: boolean
      required:
        - id
        - expenseId
        - userId
        - userName
        - amount
        - isPaid
    ExpenseWithSplits:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        groupName:
          type: string
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        paidByUserId:
          type: integer
        paidByUserName:
          type: string
        createdAt:
          type: string
          format: date-time
        splits:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseSplit"
      required:
        - id
        - groupId
        - groupName
        - description
        - amount
        - category
        - paidByUserId
        - paidByUserName
        - createdAt
        - splits
    CreateExpenseBody:
      type: object
      properties:
        groupId:
          type: integer
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        paidByUserId:
          type: integer
        splits:
          type: array
          items:
            type: object
            properties:
              userId:
                type: integer
              amount:
                type: number
            required:
              - userId
              - amount
      required:
        - groupId
        - description
        - amount
        - category
        - paidByUserId
        - splits
    UpdateExpenseBody:
      type: object
      properties:
        requestingUserId:
          type: integer
        description:
          type: string
        amount:
          type: number
        category:
          type: string
      required:
        - requestingUserId
    DeleteExpenseBody:
      type: object
      properties:
        requestingUserId:
          type: integer
      required:
        - requestingUserId
...
    Balance:
      type: object
      properties:
        groupId:
          type: integer
        groupName:
          type: string
        groupEmoji:
          type: string
        withUserId:
          type: integer
        withUserName:
          type: string
        withUserUpiId:
          type: ["string", "null"]
        amount:
          type: number
        direction:
          type: string
          enum: [owe, owed]
      required:
        - groupId
        - groupName
        - groupEmoji
        - withUserId
        - withUserName
        - amount
        - direction
    DashboardSummary:
      type: object
      properties:
        totalOwed:
          type: number
        totalOwe:
          type: number
        netBalance:
          type: number
        groupCount:
          type: integer
        recentExpenseCount:
          type: integer
        settledThisMonth:
          type: number
      required:
        - totalOwed
        - totalOwe
        - netBalance
        - groupCount
        - recentExpenseCount
        - settledThisMonth
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [expense_added, settlement_done, group_joined]
        description:
          type: string
        amount:
          type: ["number", "null"]
        groupName:
          type: string
        actorName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - description
        - groupName
        - actorName
        - createdAt