# lib/api-zod/src/generated/types/balance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Expense Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { BalanceDirection } from "./balanceDirection";

export interface Balance {
  groupId: number;
  groupName: string;
  groupEmoji: string;
  withUserId: number;
  withUserName: string;
  /** @nullable */
  withUserUpiId?: string | null;
  amount: number;
  direction: BalanceDirection;
}