# lib/api-zod/src/generated/types/createExpenseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Expense Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExpenseBodySplitsItem } from "./createExpenseBodySplitsItem";

export interface CreateExpenseBody {
  groupId: number;
  description: string;
  amount: number;
  category: string;
  paidByUserId: number;
  splits: CreateExpenseBodySplitsItem[];
}