# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Expense Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalOwed: number;
  totalOwe: number;
  netBalance: number;
  groupCount: number;
  recentExpenseCount: number;
  settledThisMonth: number;
}