# lib/api-zod/src/generated/types/expenseSplit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Expense Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseSplit {
  id: number;
  expenseId: number;
  userId: number;
  userName: string;
  amount: number;
  isPaid: boolean;
}