# lib/api-zod/src/generated/types/expenseWithSplits.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pulse Expense Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseSplit } from "./expenseSplit";

export interface ExpenseWithSplits {
  id: number;
  groupId: number;
  groupName: string;
  description: string;
  amount: number;
  category: string;
  paidByUserId: number;
  paidByUserName: string;
  createdAt: Date;
  splits: ExpenseSplit[];
}